Added Bronze and Steel Weapons
Added Bronze and Steel Weapons
-------------------------------------
If the mod is present, they should default to the Simple Ores add-on
'Fusion'. If it isn't present they will fall back to Thermal Foundation
stats for Bronze, and Railcraft for Steel.

Added Bronze and Steel Ingot - Uncraftable. Added incase the tools are
enabled without a mod that adds Bronze/Steel

Improved Anvil Repair Logic to use the Ore Dictionary.